+++ maven/datasets/general_election/__init__.py
@@ -1,5 +1,11 @@
 from .uk_2010_results import UK2010Results
+from .uk_2015_model import UK2015Model
 from .uk_2015_results import UK2015Results
 from .uk_polls import UKPolls
 
-__all__ = ['UK2010Results', 'UK2015Results', 'UKPolls']
+__all__ = [
+    'UK2010Results',
+    'UK2015Model',
+    'UK2015Results',
+    'UKPolls'
+]