+++ CHANGELOG.md
@@ -5,9 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## [Unreleased]
+<!-- ## [Unreleased] -->
+
+## [0.0.6] - 2019-07-13
 ### Added
-- Opinion polling data for UK General Elections.
+- `general-election/UK/2015/model`: model-ready datasets for the 2015/2017 UK General Elections.
+- Regional polling datasets.
 
 ## [0.0.5] - 2019-07-13
 ### Added